+++ src/shared.h
@@ -20,7 +20,7 @@
 
 #define GETV(x, i)      x[i % x.length()]    // wrapped indexing of vector
 #define GETM(x, i, j)   x(i % x.nrow(), j)   // wrapped indexing of matrix
-#define TO_INT(x)       static_cast<int>(x)
+#define TO_INT(x)       static_cast<unsigned long int>(x)
 #define TO_DBL(x)       static_cast<double>(x)
 #define VALID_PROB(p)   ((p >= 0.0) && (p <= 1.0))
